feat(ui): hook up login endpoint with UI (#72)
* refactor generate_response method

* implement basic login logic on ui
