+++ lambdas/authentication/utils/response.py
@@ -1,3 +1,5 @@
+import datetime
+import json
 from typing import Any, Dict
 
 
@@ -22,3 +24,20 @@ def generate_response(status_code: int, body: Any) -> Dict[str, Any]:
         "body": body,
         "headers": COMMON_HEADERS
     }
+
+
+def generate_success_response(access_token):
+    """
+    Generates a success response with the access token.
+
+    :param access_token: The generated access token.
+    :return: A success response containing the access token and current timestamp.
+    """
+    body = {
+        'status': 'success',
+        'payload': {
+            'access_token': access_token,
+            'created_at': str(datetime.datetime.now())
+        }
+    }
+    return generate_response(status_code=200, body=json.dumps(body))
